+++ src/components/consensus/valuation/Excel.jsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useRef } from "react";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import { SpreadSheets } from "@mescius/spread-sheets-react";
 import * as GC from "@mescius/spread-sheets";
 import * as ExcelIO from "@grapecity/spread-excelio";
@@ -6,9 +6,78 @@ import "@mescius/spread-sheets-charts";
 import "@mescius/spread-sheets-shapes";
 import ExcelJS from "exceljs";
 import { saveAs } from "file-saver";
+import { useExcelContext } from "../../../routes/valuationCreateExcel/page";
 
-export default function Excel({ formula = "", setFormula, fileInput }) {
-  const spreadRef = useRef(null);
+export default function Excel() {
+  const { formula, setFormula, fileInput, spreadRef } = useExcelContext();
+  const [excelValues, setExcelValues] = useState({
+    202312: {
+      ep: "15243",
+      el: "b",
+      se: "c",
+      fl: "d",
+      fp: "e",
+      ite: "f",
+      sr: "g",
+      oi: "h",
+      ibt: "i",
+      year: "2023",
+      rr: "j",
+      dr: "k",
+      evebitda: "l",
+      roe: "m",
+      bps: "n",
+      ni: "o",
+      fpl: "e",
+      epl: "w",
+      ts: "q",
+      stock_id: "054901",
+    },
+    202212: {
+      ep: "1",
+      el: "2",
+      se: "e",
+      fl: "d",
+      fp: "1",
+      ite: "1",
+      sr: "3",
+      oi: "3",
+      ibt: "5",
+      year: "2022",
+      rr: "3",
+      dr: "ddd",
+      evebitda: "12",
+      roe: "ef",
+      bps: "3",
+      ni: "w",
+      fpl: "f",
+      epl: "f",
+      ts: "w",
+      stock_id: "054901",
+    },
+    202112: {
+      ep: "e",
+      el: "f",
+      se: "f",
+      fl: "e",
+      fp: "e",
+      ite: "1",
+      sr: "1",
+      oi: "2",
+      ibt: "4",
+      year: "2021",
+      rr: "t",
+      dr: "t",
+      evebitda: "c",
+      roe: "e",
+      bps: "w",
+      ni: "w",
+      fpl: "w",
+      epl: "1",
+      ts: "1",
+      stock_id: "054901",
+    },
+  });
 
   useEffect(() => {
     const sheet = spreadRef.current?.getActiveSheet();
@@ -22,6 +91,20 @@ export default function Excel({ formula = "", setFormula, fileInput }) {
     spreadRef.current = spread;
     spread.setSheetCount(2);
     const sheet = spread.getActiveSheet();
+
+    // 템플릿 데이터 세팅
+    const years = Object.keys(excelValues);
+    years.forEach((year, colIndex) => {
+      sheet.setValue(0, colIndex + 1, year);
+    });
+
+    const fields = Object.keys(excelValues[years[0]]);
+    fields.forEach((field, rowIndex) => {
+      sheet.setValue(rowIndex + 1, 0, field);
+      years.forEach((year, colIndex) => {
+        sheet.setValue(rowIndex + 1, colIndex + 1, excelValues[year][field]);
+      });
+    });
 
     sheet.bind(GC.Spread.Sheets.Events.SelectionChanged, (e, info) => {
       const { row, col } = info.newSelections[0];
@@ -44,65 +127,6 @@ export default function Excel({ formula = "", setFormula, fileInput }) {
     });
   }, []);
 
-  const handleSaveExcel = () => {
-    const json = spreadRef.current.toJSON();
-    const excelIO = new ExcelIO.IO();
-    excelIO.save(
-      json,
-      (blob) => {
-        saveAs(blob, "spreadsheet.xlsx");
-      },
-      (error) => {
-        console.error(error);
-      }
-    );
-  };
-
-  const handleOpenExcel = async () => {
-    const file = fileInput.current.files[0];
-    if (file) {
-      const reader = new FileReader();
-      reader.onload = async (e) => {
-        const data = new Uint8Array(e.target.result);
-        const workbook = new ExcelJS.Workbook();
-        await workbook.xlsx.load(data);
-
-        const worksheet = workbook.getWorksheet(1); // 첫 번째 시트를 가져옴
-
-        worksheet.eachRow((row, rowNumber) => {
-          row.eachCell((cell, colNumber) => {
-            const sheet = spreadRef.current.getSheet(0);
-            if (cell.formula) {
-              sheet.setFormula(rowNumber - 1, colNumber - 1, cell.formula);
-            } else {
-              sheet.setValue(rowNumber - 1, colNumber - 1, cell.value);
-            }
-          });
-        });
-      };
-
-      reader.readAsArrayBuffer(file);
-    }
-  };
-
-  const handleAddData = () => {
-    const sheet = spreadRef.current.getActiveSheet();
-    sheet.setValue(0, 0, "New Data");
-  };
-
-  const handleGetFormula = () => {
-    const sheet = spreadRef.current.getActiveSheet();
-    const selection = sheet.getSelections()[0];
-    const row = selection.row;
-    const col = selection.col;
-    const formula = sheet.getFormula(row, col);
-    if (formula) {
-      alert(`Formula at (${row}, ${col}): ${formula}`);
-    } else {
-      alert(`No formula at (${row}, ${col})`);
-    }
-  };
-
   return (
     <div className="w-full">
       <SpreadSheets
@@ -114,30 +138,6 @@ export default function Excel({ formula = "", setFormula, fileInput }) {
           color: "#3b3b3b",
         }}
       ></SpreadSheets>
-      <div className="w-full my-4 p-4 bg-tuatara-900 rounded">
-        <p className="mb-2">내 PC에서 엑셀 가져오기 (.xlsx)</p>
-        <input type="file" ref={fileInput} />
-        <button
-          className="p-2 bg-blue-500 text-white rounded"
-          onClick={handleOpenExcel}
-        >
-          OK
-        </button>
-        <div className="">
-          <button
-            className="p-2 mr-4 bg-green-500 text-white rounded"
-            onClick={handleSaveExcel}
-          >
-            Save Excel
-          </button>
-          <button
-            className=" p-2 bg-red-500 text-white rounded"
-            onClick={handleGetFormula}
-          >
-            Get Formula
-          </button>
-        </div>
-      </div>
     </div>
   );
 }
